Adiciona raspador base para Siganet
Raspador para Balsas (MA) foi removido por ser incompatível
(publica diários da federaçao de municípios). Entretanto, o código
foi utilizado para criar o raspador base.
